IRIS FLOWER CLASSIFICATION
Iris flower classification is a classic and fundamental machine learning task that aims to categorize iris flowers into three different species based on specific characteristics of their petals and sepals. The primary objective of this task is to build a predictive model that can accurately identify and differentiate between these three species: setosa, versicolor, and virginica.